arm: uaccess: use the same segment_eq() for (NO)MMU
segment_eq() is usually used to test if get_fs() == KERNEL_DS.
On NOMMU, get_fs() is defined to KERNEL_DS and user_addr_max() is
defined as such:

segment_eq(get_fs(), USER_DS) ? TASK_SIZE : ~0UL)

This means all calls to user_addr_max() are validated against TASK_SIZE
even if they originate in the kernel. This can cause validation of
syscall parameters to fail for init's in-kernel syscalls such as
sys_open of /dev/console. Currently this works because TASK_SIZE is
defined to the size of RAM, but that only works if RAM is based at 0.
This formalizes the current policy that there is no distinction between
USER_DS and KERNEL_DS on NOMMU.