pj4b: armadaxp: enable snoop filter
This seems to fix the coherency issues in atomic ops on Armada XP,
possibly more reliably than the DSBs around ldrex/strex. It should
certainly be more performant than that approach as well. Marvell has not
explained why or if this is needed for correctness, but they were
surprised that it was disabled and indicated that that is not an
expected configuration.

This is implemented in a hacky way that's platform-specific, but good
enough for testing the behavior in a release branch.